backend/requirements.txt
langchain>=0.1.13
langchain-core>=0.1.21
langchain-google-genai>=0.0.7
langchain-community>=0.0.32
google-generativeai>=0.3.2
pinecone-client>=3.0.0
pinecone==6.0.2
google
langchain-pinecone>=0.0.3
tavily-python>=0.3.1
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1